Fixed GCC warnings.

+++ src/vulkan/vulkan-buffer.cpp
@@ -156,7 +156,7 @@ namespace nvrhi::vulkan
 #ifdef _WIN32
                 buffer->sharedHandle = m_Context.device.getMemoryWin32HandleKHR({ buffer->memory, vk::ExternalMemoryHandleTypeFlagBits::eOpaqueWin32 });
 #else
-                buffer->sharedHandle = (void*)m_Context.device.getMemoryFdKHR({ buffer->memory, vk::ExternalMemoryHandleTypeFlagBits::eOpaqueFd });
+                buffer->sharedHandle = (void*)(size_t)m_Context.device.getMemoryFdKHR({ buffer->memory, vk::ExternalMemoryHandleTypeFlagBits::eOpaqueFd });
 #endif
             }
         }
